Fixed: Improvement on project's structure
